+++ src/lib/libast/sfio/_sfputu.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -24,12 +24,12 @@
 **	Written by Kiem-Phong Vo.
 */
 
-int _sfputu(Sfio_t*	f,	/* write a portable ulong to this stream */
-	    Sfulong_t	v)	/* the unsigned value to be written */
+ptrdiff_t _sfputu(Sfio_t*	f,	/* write a portable ulong to this stream */
+		  Sfulong_t	v)	/* the unsigned value to be written */
 {
 #define N_ARRAY		(2*sizeof(Sfulong_t))
 	uchar	*s, *ps;
-	ssize_t	n, p;
+	ptrdiff_t n, p;
 	uchar		c[N_ARRAY];
 
 	if(!f || (f->mode != SFIO_WRITE && _sfmode(f,SFIO_WRITE,0) < 0))
@@ -44,7 +44,7 @@ int _sfputu(Sfio_t*	f,	/* write a portable ulong to this stream */
 	n = (ps-s)+1;
 
 	if(n > 8 || SFWPEEK(f,ps,p) < n)
-		n = SFWRITE(f,s,n); /* write the hard way */
+		n = SFWRITE(f,s,(size_t)n); /* write the hard way */
 	else
 	{	switch(n)
 		{
@@ -68,5 +68,5 @@ int _sfputu(Sfio_t*	f,	/* write a portable ulong to this stream */
 	}
 
 	SFOPEN(f,0);
-	return (int)n;
+	return n;
 }
