+++ src/lib/libast/sfio/sfpopen.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -31,7 +31,7 @@ Sfio_t*	sfpopen(Sfio_t*		f,
 {
 	Proc_t*		proc;
 	int		sflags;
-	long		flags;
+	int		flags;
 	int		pflags;
 	char*		av[4];
